session: Move token computation out of session.rs
In effort to combat overloading Session with nonrelated functionality,
token computation functions are moved out from session.rs. Specifically,
`calculate_token()` is moved to prepared.rs and
`calculate_token_for_partition_key()` is moved to partitioner.rs.
